// Front-End/src/app/models/client.company.region.ts
import { StatusEnabDisabEnum } from "./status-enab-disab-enum";

export class ClientCompanyRegion {
    companyId: number | null = null;
    resaleId: number | null = null;
    id: number | null = null;
    status: StatusEnabDisabEnum = StatusEnabDisabEnum.DISABLED;
    description: string = '';
    uf: string = '';
}